+++ lib/eve_online.rb
@@ -85,7 +85,10 @@ require 'eve_online/esi/character_loyalty_points'
 require 'eve_online/esi/character_wallet'
 require 'eve_online/esi/character_assets'
 require 'eve_online/esi/character_attributes'
+# require 'eve_online/esi/character_clones'
 require 'eve_online/esi/server_status'
+require 'eve_online/esi/races'
+require 'eve_online/esi/bloodlines'
 
 require 'eve_online/esi/models/base'
 require 'eve_online/esi/models/loyalty_point'
@@ -94,6 +97,9 @@ require 'eve_online/esi/models/skill'
 require 'eve_online/esi/models/asset'
 require 'eve_online/esi/models/attributes'
 require 'eve_online/esi/models/character'
+
+require 'eve_online/esi/models/race'
+require 'eve_online/esi/models/bloodline'
 
 # SDE
 require 'eve_online/sde/base'